requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
chardet==3.0.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
yarl==1.5.1
